% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.r
\name{mult}
\alias{mult}
\title{multipy together two numbers}
\usage{
mult(4,5)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}

\item{product}{The product of multiplying x and y}
}
\description{
multipy together two numbers
}
